removed unused imports and fixed bug in dmc_dreamer exp
