+++ rlkit/torch/model_based/dreamer/experiments/dmc_dreamer_experiment.py
@@ -1,143 +1,19 @@
-import gym
-import numpy as np
-
-
-class DeepMindControl:
-    def __init__(self, name, size=(64, 64), camera=None):
-        domain, task = name.split("_", 1)
-        if domain == "cup":
-            domain = "ball_in_cup"
-        if isinstance(domain, str):
-            from dm_control import suite
-
-            self._env = suite.load(domain, task)
-        else:
-            assert task is None
-            self._env = domain()
-        self._size = size
-        if camera is None:
-            camera = dict(quadruped=2).get(domain, 0)
-        self._camera = camera
-
-    @property
-    def observation_space(self):
-
-        spaces = {}
-        for key, value in self._env.observation_spec().items():
-            spaces[key] = gym.spaces.Box(-np.inf, np.inf, value.shape, dtype=np.float32)
-        spaces["image"] = gym.spaces.Box(0, 255, self._size + (3,), dtype=np.uint8)
-        return gym.spaces.Dict(spaces)
-
-    @property
-    def action_space(self):
-        spec = self._env.action_spec()
-        return gym.spaces.Box(spec.minimum, spec.maximum, dtype=np.float32)
-
-    def step(self, action):
-        time_step = self._env.step(action)
-        obs = dict(time_step.observation)
-        obs["image"] = self.render()
-        reward = time_step.reward or 0
-        done = time_step.last()
-        info = {"discount": np.array(time_step.discount, np.float32)}
-        return obs, reward, done, info
-
-    def reset(self):
-        time_step = self._env.reset()
-        obs = dict(time_step.observation)
-        obs["image"] = self.render()
-        return obs
-
-    def render(self, *args, **kwargs):
-        if kwargs.get("mode", "rgb_array") != "rgb_array":
-            raise ValueError("Only render mode 'rgb_array' is supported.")
-        return self._env.physics.render(*self._size, camera_id=self._camera)
-
-
-class TimeLimit:
-    def __init__(self, env, duration):
-        self._env = env
-        self._duration = duration
-        self._step = None
-
-    def __getattr__(self, name):
-        return getattr(self._env, name)
-
-    def step(self, action):
-        assert self._step is not None, "Must reset environment."
-        obs, reward, done, info = self._env.step(action)
-        self._step += 1
-        if self._step >= self._duration:
-            done = True
-            if "discount" not in info:
-                info["discount"] = np.array(1.0).astype(np.float32)
-            self._step = None
-        return obs, reward, done, info
-
-    def reset(self):
-        self._step = 0
-        return self._env.reset()
-
-
-class ActionRepeat:
-    def __init__(self, env, amount):
-        self._env = env
-        self._amount = amount
-
-    def __getattr__(self, name):
-        return getattr(self._env, name)
-
-    def step(self, action):
-        done = False
-        total_reward = 0
-        current_step = 0
-        while current_step < self._amount and not done:
-            obs, reward, done, info = self._env.step(action)
-            total_reward += reward
-            current_step += 1
-        return obs, total_reward, done, info
-
-
-class NormalizeActions:
-    def __init__(self, env):
-        self._env = env
-        self._mask = np.logical_and(
-            np.isfinite(env.action_space.low), np.isfinite(env.action_space.high)
-        )
-        self._low = np.where(self._mask, env.action_space.low, -1)
-        self._high = np.where(self._mask, env.action_space.high, 1)
-
-    def __getattr__(self, name):
-        return getattr(self._env, name)
-
-    @property
-    def action_space(self):
-        low = np.where(self._mask, -np.ones_like(self._low), self._low)
-        high = np.where(self._mask, np.ones_like(self._low), self._high)
-        return gym.spaces.Box(low, high, dtype=np.float32)
-
-    @property
-    def observation_space(self):
-        return gym.spaces.Box(0, 255, (64 * 64 * 3,), dtype=np.uint8)
-
-    def step(self, action):
-        original = (action + 1) / 2 * (self._high - self._low) + self._low
-        original = np.where(self._mask, original, action)
-        o, r, d, i = self._env.step(original)
-        o = o["image"].transpose(2, 0, 1).flatten()
-        return o, r, d, i
-
-    def reset(self):
-        return self._env.reset()["image"].transpose(2, 0, 1).flatten()
-
-
 def experiment(variant):
     import os
 
-    from rlkit.torch.model_based.dreamer.kitchen_video_func import video_post_epoch_func
+    from rlkit.envs.wrappers.dmc_wrappers import (
+        ActionRepeat,
+        DeepMindControl,
+        NormalizeActions,
+        TimeLimit,
+    )
+    from rlkit.torch.model_based.dreamer.kitchen_video_func import (
+        video_post_epoch_func_,
+    )
 
     os.environ["D4RL_SUPPRESS_IMPORT_ERROR"] = "1"
     import torch
+    import torch.nn as nn
 
     import rlkit.torch.pytorch_util as ptu
     from rlkit.envs.mujoco_vec_wrappers import DummyVecEnv
@@ -181,7 +57,7 @@ def experiment(variant):
     actor = ActorModel(
         variant["model_kwargs"]["model_hidden_size"],
         world_model.feature_size,
-        hidden_activation=torch.nn.functional.elu,
+        hidden_activation=nn.ELU,
         discrete_action_dim=0,
         continuous_action_dim=eval_env.action_space.low.size,
         **variant["actor_kwargs"],
@@ -191,14 +67,14 @@ def experiment(variant):
         * variant["vf_kwargs"]["num_layers"],
         output_size=1,
         input_size=world_model.feature_size,
-        hidden_activation=torch.nn.functional.elu,
+        hidden_activation=nn.ELU,
     )
     target_vf = Mlp(
         hidden_sizes=[variant["model_kwargs"]["model_hidden_size"]]
         * variant["vf_kwargs"]["num_layers"],
         output_size=1,
         input_size=world_model.feature_size,
-        hidden_activation=torch.nn.functional.elu,
+        hidden_activation=nn.ELU,
     )
     variant["trainer_kwargs"]["target_vf"] = target_vf
 
@@ -276,10 +152,10 @@ def experiment(variant):
     )
     algorithm.to(ptu.device)
     if variant.get("save_video", False):
-        algorithm.post_epoch_funcs.append(video_post_epoch_func)
+        algorithm.post_epoch_funcs.append(video_post_epoch_func_)
     print("TRAINING")
     algorithm.to(ptu.device)
     algorithm.train()
     if variant.get("save_video", False):
-        video_post_epoch_func(algorithm, -1)
+        video_post_epoch_func_(algorithm, -1)
     algorithm.train()
