+++ rlkit/torch/model_based/dreamer/train_world_model.py
@@ -1,5 +1,4 @@
 import os
-import os.path as osp
 
 import cv2
 import h5py
